# Improve `factorial(n, k)` performance for `BigInt`